+++ fittrackee_client/src/components/Activities/ActivitiesList.jsx
@@ -2,7 +2,7 @@ import { format } from 'date-fns'
 import React from 'react'
 import { Link } from 'react-router-dom'
 
-import { getDateWithTZ } from '../../utils'
+import { apiUrl, getDateWithTZ } from '../../utils'
 
 export default class ActivitiesList extends React.PureComponent {
   render() {
@@ -36,10 +36,34 @@ export default class ActivitiesList extends React.PureComponent {
                         alt="activity sport logo"
                       />
                     </td>
-                    <td>
+                    <td className="activity-title">
                       <Link to={`/activities/${activity.id}`}>
                         {activity.title}
                       </Link>
+                      {activity.map && (
+                        <span>
+                          <img
+                            className="activity-map"
+                            src={`${apiUrl}activities/map/${
+                              activity.map
+                            }?${Date.now()}`}
+                            alt="activity map"
+                          />
+                          <span className="map-attribution text-right">
+                            <div>
+                              <span className="map-attribution-text">©</span>
+                              <a
+                                className="map-attribution-text"
+                                href="http://www.openstreetmap.org/copyright"
+                                target="_blank"
+                                rel="noopener noreferrer"
+                              >
+                                OpenStreetMap
+                              </a>
+                            </div>
+                          </span>
+                        </span>
+                      )}
                     </td>
                     <td>
                       {format(